addition of tests
added each input file and output file for every test. added an excel spreadsheet for list of test cases, will add it to the pdf once we are done.

I'm not quite sure about the session summary assoicated with logout. i probably should make the summary as complicated as possible, but how the other functions behave should not be under logout. Will change once decided.